# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estacionamento API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Buscar todas os funcionários",
                "operationId": "e7707cd772b42c62a4f8dcec4eac24dd",
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EmployeesResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Criar novo funcionário",
                "operationId": "8feefa7a242d755fa119bf6a4d8ca6f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmployeesDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeesResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employees/{parking_id}/{employees_id}": {
            "get": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Buscar funcionário por ID do estacionamento e ID do funcionário",
                "operationId": "48d4d5f395f6030705814c50aebf4c86",
                "parameters": [
                    {
                        "name": "parking_id",
                        "in": "path",
                        "description": "ID do estacionamento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "employees_id",
                        "in": "path",
                        "description": "ID do Funcionário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeesResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Deletar funcionário por ID do estacionamento e ID do funcionário",
                "operationId": "4994cac8fb13fdb5cd0c89dbc007ab09",
                "parameters": [
                    {
                        "name": "parking_id",
                        "in": "path",
                        "description": "ID do estacionamento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "employees_id",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Funcionários"
                ],
                "summary": "Atualizar funcionário por ID do estacionamento e ID do funcionário",
                "operationId": "c0b6fa5f3f63b72ad1f6e4db7e74d03a",
                "parameters": [
                    {
                        "name": "parking_id",
                        "in": "path",
                        "description": "ID do estacionamento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "employees_id",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmployeesDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta bem sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeesResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmployeesDTO": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Pedro"
                    },
                    "cpf": {
                        "type": "string",
                        "example": "123123123"
                    },
                    "email": {
                        "type": "string",
                        "example": "exemplo@exemplo.com"
                    },
                    "office": {
                        "type": "string",
                        "example": "CEO"
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    },
                    "parking_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "EmployeesResource": {
                "properties": {
                    "errors": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "content": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "nome": {
                                "type": "string"
                            },
                            "cpf": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "cargo": {
                                "type": "string"
                            },
                            "ativo": {
                                "type": "boolean"
                            },
                            "estacionamento_id": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Autenticação",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}